+++ Sources/SuperwallKit/Web/WebEntitlementRedeemer.swift
@@ -147,6 +147,10 @@ actor WebEntitlementRedeemer {
 
       switch type {
       case .code:
+        await MainActor.run {
+          superwall.paywallViewController?.loadingState = .manualLoading
+          superwall.paywallViewController?.closeSafari()
+        }
         await delegate.willRedeemLink()
       case .existingCodes:
         break
@@ -208,6 +212,9 @@ actor WebEntitlementRedeemer {
       // Call the delegate if user try to redeem a code
       if case let .code(code) = type {
         if let codeResult = response.results.first(where: { $0.code == code }) {
+          await MainActor.run {
+            superwall.paywallViewController?.loadingState = .ready
+          }
           await delegate.didRedeemLink(result: codeResult)
         }
       }
@@ -237,6 +244,9 @@ actor WebEntitlementRedeemer {
         )
         redemptions.append(errorResult)
 
+        await MainActor.run {
+          superwall.paywallViewController?.loadingState = .ready
+        }
         await delegate.didRedeemLink(result: errorResult)
       }